Optimise construction a bit
Make tables of fixed values compute only once, on first construction
Put those tables in RAM2, which is less-used on Teensy than RAM1, but just as fast

First construction takes 1400us, subsequent ones about 6us. This is an improvement.

TODO: tables should maybe be malloc'ed, so if we have no instances of AudioSynthDexed in memory we don't waste 32kB of RAM